Add defaults class and rename ingest_data.
This commit adds a defaults class in which all the default
values for each of the methods are defined. This way we don't
need to pass always all the arguments when importing some of
the managers functions and at the same time keep all defaults
in a central place.

Also, it renames de ingest_data function into create_data to
keep consistent with the rest of create functions.