Add support for RBAC in weaviate-cli.
This commit adds the option to add, remove, get roles.
Assigning roles to users, as well as permissions to roles,
revoking them (roles as well as permissions).

It also adds the option to specify a user from your
config as the user to pass in the client connection. This
allows having multiple api-keys assigned to different users
without having the need to keep different configs.