Improve the randomized generated data.
Rely on the Faker library to generate relevant strings
which can then be vectorized.
This PR improvers the way the data is vectorized, ressambling
the content from the movies.json dataset and allowing adding
several millions of objects.
Also, it uses the multi-processing functionality from python
to build big datasets in parallel.
Adds a --verbose during data creation, update and delete to
track the progress.
It also adds an offset when performing updates in very big
collections, so that the updates do not occur always on the
first items.