Add support for named vectors.
Up to know there was no proper support of named vectors, it
could be configured, but quite ad-hoc. This PR adds the posibility
to configure named vectors by passing the flag --named_vector
during creation, otherwise the vectorizer will be the common one.

The get collections has also been modified to display if the
collection is configured with named vectors or not, displaying the
corresponding vectorizer and vector index for the named vector.

Last, but not least, some tests to verify that all this logic
isn't impacted by future changes have been added.