Support using alias to ingest data.
There were few checks verifying that the collection
would exist before ingesting/updating/deleting data.
However, if an alias exists it is not found when listing
collections. Therefore, a better check had to be included,
as well as some code to use the collection with the alias
right before the data operation is about to take place (it's
not possible to do collection modifications or getting config
using an alias)