Add file-alias and fail-on-timeout.
One issue we have, is if launching several QPS benchmarks
at the same time, the name will be the same. Also, sometimes
it's hard to differentiate from which run each file is. For
that reason we are adding a new --file-alias flag that
allows us passing some identifier to have an identified
run.

Also, this PR adds a new flag --fail-on-timeout to ensure
the QPS benchmark errors and exists the benchmark once
some of the queries takes longer than the defined 10s
timeout.